fixed some stuff, broke some stuff

+++ Assignment1/unitConversion.cpp
@@ -1,12 +1,12 @@
-int unitConversion(int temperature, char units) {
+double unitConversion(double temperature, char units) {
     // main math calculation 
-    int calculation;
+    double calculation;
 
     if (units == 'C') {
-        calculation = (temperature * 9 / 5) + 32 ;
+        calculation = (temperature * 9 / 5) + 32;
     }
     else {
-        calculation = (temperature - 32) * 5/9 ; 
+        calculation = (temperature - 32) * 5/9 ;
     }
 
     return calculation;
