+++ Assignment1/unitConversionHeader.h
@@ -2,6 +2,6 @@
 #ifndef MYFUNCTIONS_H
 #define MYFUNCTIONS_H
 
-extern int unitConversion(int, char);  // Function declaration
+extern double unitConversion(double, char);  // Function declaration
 
 #endif
